requirements.txt
aerich==0.7.1
aiosqlite==0.17.0
anyio==3.7.1
asgiref==3.7.2
asyncpg==0.28.0
atlastk==0.13.2
certifi==2023.5.7
click==8.1.5
dictdiffer==0.9.0
dnspython==2.4.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.78.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
idna==3.4
iso8601==1.1.0
Mako==1.2.4
MarkupSafe==2.1.3
pydantic==1.10.11
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.19.1
tomlkit==0.11.8
tortoise==0.1.1
tortoise-orm==0.19.3
typing_extensions==4.7.1
uvicorn==0.17.6
uvloop==0.17.0
watchgod==0.8.2
websockets==11.0.3
